Simplify unnecessarily complex logic in ask_vault_passwords()
ask_vault_passwords() was called from three places; two just wanted the
current vault password, and one wanted to set self.vault_pass most of
the time, but sometimes with a different prompt, and sometimes also to
set self.vault_new_pass, and required the caller to indicate which it
wanted by setting flags, and always returned two values, which would be
equal if a new_vault_pass wasn't required.

The new ask_vault_password() function asks for a vault password. If the
caller wants another (new) vault password with a different prompt, it
can call the function again. If it wants to set new_vault_pass, it can
call the function and assign its result to new_vault_pass. The logic
both inside the function and in the caller becomes obvious.